Support O-MR1 vendor.img with P-GSI
Include "-v" or "--vendor" option that is followed by a vendor
version which is now "8.1.0" for O-MR1.
This does not break current usage, but add an option to support
O-MR1 vendor.img.
By adding "-v 8.1.0" as a parameter, the script will modify
"ro.build.version.release" property in GSI to the specified vendor
version to match the keymaster version binding.
For O-MR1, it will add "ro.vndk.version=27" to GSI.

Bug: 72405134
Test: 1. Flash O-MR1 wahoo device.
      2. Build aosp_arm64_ab-userdebug with vndk snapshot v27.
      3. Run change_security_patch_ver.sh \
             system.img system_gsi.img -v 8.1.0
      4. Flash the device with vbmeta.img and system_gsi.img.
      5. Check boot.

Change-Id: I2376db30f34ce709fa28b05eb729268159e17f46